wip: refactor(actions): streamline action handling and introduce action categories
This commit refactors the action handling system by introducing a new `ActionCategory` enum to categorize actions such as Movement, Attack, Wait, and more. It updates the `GameAction` trait to include a `category` method, enhancing the clarity and organization of action types.

Additionally, the `TurnActor` component is modified to support a maximum number of queued actions, improving memory management. The action queueing methods are updated to ensure actions are added with respect to this limit, preventing overflow.

Various action implementations, including `Wait` and `Walk`, are adjusted to utilize the new category system, ensuring consistent behavior across the action framework. This refactor enhances maintainability and prepares the codebase for future action types and behaviors.

+++ crates/echos_assets/src/entities/components.rs
@@ -20,13 +20,13 @@ impl Default for AIBehaviorType {
 #[derive(Serialize, Deserialize, Debug, Clone, Reflect)]
 pub struct TurnActorData {
     /// Speed value for turn timing
-    pub speed: u64,
+    pub speed: u32,
     /// Maximum number of actions that can be queued
     pub action_queue_size: Option<usize>,
 }
 
 impl TurnActorData {
-    pub fn new(speed: u64) -> Self { Self { speed, action_queue_size: None } }
+    pub fn new(speed: u32) -> Self { Self { speed, action_queue_size: None } }
 
     pub fn with_queue_size(mut self, size: usize) -> Self {
         self.action_queue_size = Some(size);
